refactor(veil): #2271: move voting dialog to new directory
